Code refactoring and deleting unnecessary files
Fixed compilation problem caused by VertexTypes and CGraphicsManager.
CRenderableVertex and CIndexedVertex have been moved from CTestProcess
to their corresponding files.For last we draw a quad with a texture on
it.

+++ mvuab/Code/Videogame/VideogameProcess.cpp
@@ -83,6 +83,6 @@ void CVideogameProcess::Render()
 	if( pInputManager->IsDown( IDV_MOUSE, MOUSE_BUTTON_LEFT ) )
 	{
 		Vect3i delta = pInputManager->GetMouseDelta();
-		m_pCamera->AddYawPitch(delta.x,delta.y);
+		m_pCamera->AddYawPitch((float) delta.x, (float) delta.y);
 	}
 }
